+++ packages/black-forest-labs/CHANGELOG.md
@@ -1,5 +1,23 @@
 # @ai-sdk/black-forest-labs
 
+## 0.0.5
+
+### Patch Changes
+
+- 473a3f5: feat(provider/black-forest-labs): Add new provider options
+
+## 0.0.4
+
+### Patch Changes
+
+- 539e89d: fix (provider/black-forest-labs): allow null for cost and megapixel in provider response
+
+## 0.0.3
+
+### Patch Changes
+
+- 7f40d34: feat (provider/black-forest-labs): include cost and megapixels in metadata
+
 ## 0.0.2
 
 ### Patch Changes