+++ packages/google-vertex/CHANGELOG.md
@@ -1,5 +1,19 @@
 # @ai-sdk/google-vertex
 
+## 3.0.79
+
+### Patch Changes
+
+- Updated dependencies [dcdfeb7]
+  - @ai-sdk/anthropic@2.0.49
+
+## 3.0.78
+
+### Patch Changes
+
+- Updated dependencies [f2c6cbd]
+  - @ai-sdk/anthropic@2.0.48
+
 ## 3.0.77
 
 ### Patch Changes